+++ changelog.md
@@ -4,7 +4,7 @@
 - Added Headers to each bonus section on item sheet
 - Added Target Toggle to switch between any/all targets.
 - Better support for items in containers 
-  - Primarily when selecting Item/Weapon targets and when choosing what options are available for Weapon Type targets
+  - Primarily when selecting Item/Weapon targets and when choosing what options are available for Weapon Type and custom Weapon Group targets
 
 ### Bugfixes
 - Fixed remaining warning for rolls